+++ Cargo.toml
@@ -23,6 +23,7 @@ serde = { version = "1.0", optional = true, features = ["derive"] }
 fff = { version = "0.2.0", features = ["derive"] }
 groupy = "0.3.0"
 rand_core = "0.5"
+lazy_static = "1.4.0"
 
 [dev-dependencies]
 rand_xorshift = "0.2"
